added NextBumpTime

+++ src/discord/bot/gq/event/InviteLinkDeletion.java
@@ -11,10 +11,12 @@ public class InviteLinkDeletion extends ListenerAdapter {
 
         String userMessage = event.getMessage().getContentRaw();
         long authorId = event.getAuthor().getIdLong();
+        long channelId = event.getChannel().getIdLong();
         long ownerId = 739143338975952959L;
         Member author = event.getMember();
         assert author != null;
         boolean isStaff = author.hasPermission(Permission.MESSAGE_MANAGE);
+        long yourProjectsChannelId = 784973821790257192L;
 
 
         if (authorId == ownerId) {
@@ -25,12 +27,15 @@ public class InviteLinkDeletion extends ListenerAdapter {
             return;
         }
 
+        if (channelId == yourProjectsChannelId) {
+            return;
+        }
+
         if (!userMessage.contains("https://discord.gg")) {
             return;
         }
 
         event.getMessage().delete().queue();
         event.getChannel().sendMessage("Hier dürfen keine Invitelinks gepostet werden " + event.getAuthor().getAsMention() + "!").queue();
-
     }
 }
